// server/src/types/Game.ts
export enum GAME_STATUS {
    PENDING = "pending",
    LOSE = "lose",
    WIN = "win",
}

export interface IGameRenderData {
    players: IPlayerRenderData[];
    localPlayer: IPlayerRenderData;
    tileService: ITilesServiceRenderData;
    characterService: ICharacterServiceRenderData;
    resourceService: IResourceServiceRenderData;
    constructionService: IConstructionServiceRenderData;
    inventionService: IInventionServiceRenderData;
    eventService: IEventServiceRenderData;
    equipmentService: IEquipmentRenderData;
    arrangeCampRestService: IArrangeCampRestServiceRenderData;
    beastService: IBeastServiceRenderData;
    phaseService: IPhaseServiceRenderData;
    moraleService: IMoraleRenderData;
    round: number;
    logs: ILogMessageRenderData[];
    actionService: IActionServiceRenderData;
    alertService: IAlertServiceRenderData;
    scenarioService: IScenarioServiceRenderData;
    weatherService: IWeatherServiceRenderData;
    allPawns: IPawnRenderData<any>[];
    tokenService: ITokenServiceRenderData;
    adventureService: IAdventureServiceRenderData;
    mysteryService: IMysteryServiceRenderData;
}

export interface IGame {
    playerService: IPlayerService;
    localPlayer: Player;
    characterService: ICharacterService;
    tileService: ITileService;
    resourceService: IResourceService;
    constructionService: IConstructionService;
    inventionService: IInventionService;
    eventService: IEventService;
    equipmentService: IEquipment;
    actionSlotService: IActionSlotService;
    arrangeCampRestService: IArrangeCampRestService;
    beastService: IBeastService;
    moraleService: IMorale;
    tokenService: ITokenService;
    weatherService: IWeatherService;
    phaseService: IPhaseService;
    actionService: IActionService;
    chatLog: IChatLog;
    alertService: IAlertService;
    scenarioService: IScenarioService;
    allPawns: IPawn[];
    round: number;
    otherPawns: IPawn[];
    setNextRound: () => void;
    setPawn: (droppableId: string, draggableId: string) => void;

    gameStatus: GAME_STATUS;
    unsetPawn: (droppableId: string, draggableId: string) => void;

    canCommitResources: (droppableId: string) => boolean;
    shouldCommitResources: (droppableId: string) => boolean;
    resetPawns: () => void;
    actionSlotRenderData: IActionSlotServiceRenderData;
    adventureService: IAdventureService;
    mysteryService: IMysteryService;

    addToOtherPawns: (pawn: IPawn[] | IPawn) => void;

    setGameStatus: (status: GAME_STATUS.WIN | GAME_STATUS.LOSE, reason?: string) => void;
    renderData: IGameRenderData;
}